working API call to holiday
