+++ src/pages/HolidayShow.js
@@ -4,18 +4,16 @@ import {Link} from "react-router-dom"
 
 
 import React, { Component } from 'react'
-import ApiModel from '../models/apimodel'
+// import products from '../models/apimodel';
 
 
 
+// const {foo, bar}  = await products.then(result => result.data);
+// console.log(foo, bar)
 
 function HolidayShow(props) {
   
-  const fetchData = () => {
-    ApiModel.all().then(data => {
-      this.setState({ api: data })
-    })
-  }
+
 
 
 
@@ -42,27 +40,43 @@ function HolidayShow(props) {
     });
   };
 
+  
+
   const loaded = () => {
-      return props.HolidayItem.map((item) => (
-          <div key={item._id}>
-              <ul>
-                  <li>{item.itemName}</li>
-                  <li>{item.itemDescription}</li>
-                  <li>{item.itemUrl}</li>
-                </ul>
-        <Link to={`/hol-registry/${item._id}`}><button>EDIT</button></Link>
-      <button id="delete" onClick={() => props.deleteHolidayItem(item._id)}>
-        DELETE
-      </button>
-          </div>
-      ))
-  }
+    return [
+      props.ApiResponse.map((products) => (
+      <div>
+        <h1>Some recommendations from Etsy!</h1>
+          <ul>
+              <li>{products.name}</li>
+              <li><img src={products.image}></img></li>
+              <li><a href={products.url}>{products.name}</a></li>
+            </ul>
+    
+      </div>
+  ))
+  , props.HolidayItem.map((item) => (
+      <div key={item._id}>
+          <ul>
+              <li>{item.itemName}</li>
+              <li>{item.itemDescription}</li>
+              <li>{item.itemUrl}</li>
+            </ul>
+    <Link to={`/hol-registry/${item._id}`}><button>EDIT</button></Link>
+  <button id="delete" onClick={() => props.deleteHolidayItem(item._id)}>
+    DELETE
+  </button>
+      </div>
+  ))]
+}
+ 
+  
   const loading = () => {
     return <h1>Loading...</h1>;
   };
   return (
     <section>
-      <div>{this}</div>
+      <div></div>
       <form onSubmit={handleSubmit}>
         <input
           type="text"
